refactor: move AllPages to common package
