+++ internal/service/streamprocessor/data_source_plural.go
@@ -2,11 +2,11 @@ package streamprocessor
 
 import (
 	"context"
-	"errors"
 	"net/http"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/dsschema"
 	"go.mongodb.org/atlas-sdk/v20240805001/admin"
 )
 
@@ -29,7 +29,7 @@ func (d *streamProcessorsDS) Read(ctx context.Context, req datasource.ReadReques
 		ItemsPerPage: conversion.Int64PtrToIntPtr(itemsPerPage),
 		PageNum:      conversion.Int64PtrToIntPtr(pageNum),
 	}
-	sdkProcessors, err := AllPages(ctx, func(ctx context.Context, pageNum int) (PaginateResponse[admin.StreamsProcessorWithStats], *http.Response, error) {
+	sdkProcessors, err := dsschema.AllPages(ctx, func(ctx context.Context, pageNum int) (dsschema.PaginateResponse[admin.StreamsProcessorWithStats], *http.Response, error) {
 		request := connV2.StreamsApi.ListStreamProcessorsWithParams(ctx, &params)
 		request = request.PageNum(pageNum)
 		return request.Execute()
@@ -45,27 +45,4 @@ func (d *streamProcessorsDS) Read(ctx context.Context, req datasource.ReadReques
 		return
 	}
 	resp.Diagnostics.Append(resp.State.Set(ctx, newStreamConnectionsModel)...)
-}
-
-type PaginateResponse[T any] interface {
-	GetResults() []T
-}
-
-func AllPages[T any](ctx context.Context, call func(ctx context.Context, pageNum int) (PaginateResponse[T], *http.Response, error)) ([]T, error) {
-	var results []T
-	for i := 1; ; i++ {
-		resp, _, err := call(ctx, i)
-		if err != nil {
-			return nil, err
-		}
-		if resp == nil {
-			return nil, errors.New("no response")
-		}
-		currentResults := resp.GetResults()
-		if len(currentResults) == 0 {
-			break
-		}
-		results = append(results, currentResults...)
-	}
-	return results, nil
 }
